+++ modules/server-node/src/services/store.ts
@@ -235,6 +235,7 @@ const convertTransferEntityToFullTransferState = (
 };
 
 const convertEntitiesToWithdrawalCommitment = (
+  transferEntity: Transfer,
   resolveEntity: Update | null,
   createEntity: Update,
   channel: Channel,
@@ -259,7 +260,7 @@ const convertEntitiesToWithdrawalCommitment = (
     nonce: initialState.nonce,
     callData: initialState.callData,
     callTo: initialState.callTo,
-    transactionHash: resolveMeta.transactionHash ?? undefined,
+    transactionHash: transferEntity.transactionHash ?? resolveMeta.transactionHash ?? undefined,
   };
 };
 
@@ -422,7 +423,7 @@ export class PrismaStore implements IServerNodeStore {
       throw new Error("Could not retrieve channel for withdraw commitment");
     }
 
-    return convertEntitiesToWithdrawalCommitment(entity.resolveUpdate!, entity.createUpdate!, channel);
+    return convertEntitiesToWithdrawalCommitment(entity, entity.resolveUpdate!, entity.createUpdate!, channel);
   }
 
   async getWithdrawalCommitment(transferId: string): Promise<WithdrawCommitmentJson | undefined> {
@@ -446,7 +447,7 @@ export class PrismaStore implements IServerNodeStore {
       throw new Error("Could not retrieve channel for withdraw commitment");
     }
 
-    return convertEntitiesToWithdrawalCommitment(entity.resolveUpdate!, entity.createUpdate!, channel);
+    return convertEntitiesToWithdrawalCommitment(entity, entity.resolveUpdate!, entity.createUpdate!, channel);
   }
 
   async saveWithdrawalCommitment(transferId: string, withdrawCommitment: WithdrawCommitmentJson): Promise<void> {
@@ -503,7 +504,7 @@ export class PrismaStore implements IServerNodeStore {
       entities
         .map((e) => {
           return {
-            commitment: convertEntitiesToWithdrawalCommitment(e.resolveUpdate, e.createUpdate!, e.channel!),
+            commitment: convertEntitiesToWithdrawalCommitment(e, e.resolveUpdate, e.createUpdate!, e.channel!),
             transfer: convertTransferEntityToFullTransferState(e),
           };
         })